# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Personal OneDrive Document Finder Agent",
    "description": "A Microsoft Copilot Studio–compatible API that lets a signed-in user search their OneDrive, retrieve document metadata, and optionally generate AI summaries **on demand** to minimise cost.\n\n**Authentication:** Bearer token (Azure AD delegated flow — `Files.Read` scope required).\n\n> ⚡ `/search` and `/document/{id}` — zero AI cost, fast responses.\n> 🧠 `/summarize` — AI called only on explicit user request.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://your-api-host.azurewebsites.net",
      "description": "Production (Azure App Service)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/document/{item_id}": {
      "get": {
        "tags": ["Document"],
        "operationId": "getDocument",
        "summary": "Get document details",
        "description": "Retrieve metadata and a short-lived download URL for a specific OneDrive document.\n\n**No AI is used. Document content is not read.**",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "description": "OneDrive item ID (obtained from /search results)",
            "schema": {
              "type": "string",
              "example": "01BYE5RZ6QN3ZWBTUCIRGZ"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document metadata returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentDetail" },
                "example": {
                  "id": "01BYE5RZ6QN3ZWBTUCIRGZ",
                  "name": "Q4 Budget Report.xlsx",
                  "webUrl": "https://onedrive.live.com/edit.aspx?...",
                  "size": 524288,
                  "createdDateTime": "2024-10-01T08:00:00Z",
                  "lastModifiedDateTime": "2024-11-15T09:23:00Z",
                  "mimeType": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Azure AD access token with `Files.Read` delegated scope. Audience: `api://<AZURE_CLIENT_ID>`"
      }
    },
    "schemas": {
...
      "DocumentDetail": {
        "type": "object",
        "required": ["id", "name", "webUrl"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "webUrl": { "type": "string", "format": "uri" },
          "size": { "type": "integer", "nullable": true, "description": "File size in bytes" },
          "createdDateTime": { "type": "string", "format": "date-time", "nullable": true },
          "lastModifiedDateTime": { "type": "string", "format": "date-time", "nullable": true },
          "mimeType": { "type": "string", "nullable": true, "description": "MIME type" },
          "@microsoft.graph.downloadUrl": {
            "type": "string",
            "nullable": true,
            "description": "Pre-authenticated download URL (valid ~1 hour)"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "description": "Short error code" },
          "message": { "type": "string", "description": "Human-readable description" },
          "details": { "type": "string", "nullable": true, "description": "Debug info (dev only)" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid access token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "unauthorized", "message": "Token has expired." }
          }
        }
      },
      "Forbidden": {
        "description": "Token lacks required permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Document not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  },
...
}